Reworking the image API to incorporate database. WIP.
